Refactor event system, add logging, and modularize GUI components.

+++ src/event/economy/ammo.rs
@@ -6,5 +6,5 @@ pub struct BuyAmmo {
     pub timestamp: String,
 
     #[serde(rename = "Cost")]
-    pub cost: u64
+    pub cost: u32
 }
